feat: First iteration for solving the "paths problem" (#3)
